[Fix] mainClub 탠스택쿼리 이주
