+++ src/pages/MainClub/index.js
@@ -1,2 +1,3 @@
 export * from './loader';
 export * from './MainClub';
+export * from './fetchClubList';
