+++ src/pages/MainClub/loader.js
@@ -1,35 +1,14 @@
-import pb from '@/api/pocketbase';
-import { getPbImgs } from '@/utils';
+import { queryClient } from '@/client/queryClient';
+import { fetchClubList } from './fetchClubList';
 
-export async function loader({ request }) {
-  const url = new URL(request.url);
-  const filters = url.searchParams.get('filters');
+export const loader =
+  () =>
+  async ({ request }) => {
+    const url = new URL(request.url);
+    const filters = url.searchParams.get('filters');
 
-  let filterQuery = null;
-
-  if (filters) {
-    if (filters.includes(',')) {
-      filterQuery = filters
-        .split(',')
-        .map((filter) => `genre.title~"${filter}"`)
-        .join('||');
-    } else {
-      filterQuery = `genre.title~"${filters}"`;
-    }
-  }
-
-  const clubs = await pb.collection('socialing').getFullList({
-    fields:
-      'id,title,dateTime,isOffline,collectionId,location,limitPerson,confirmUser,img,expand.genre.title',
-    expand: 'genre',
-    filter: filterQuery,
-    sort: '-created',
-  });
-
-  const clubItems = clubs.map((club) => {
-    const photoURL = getPbImgs(club);
-    club.photo = photoURL;
-    return club;
-  });
-  return clubItems;
-}
+    return await queryClient.ensureQueryData({
+      queryKey: ['mainClub', filters],
+      queryFn: fetchClubList(filters),
+    });
+  };
